refactor some duplicate component and action-creator logic into separate files and directories. also updated axios version to fix vulnerability
