--- packages/local-client/src/components/action-bar.tsx
@@ -1,42 +1,0 @@
-import './action-bar.css';
-import { useActions } from '../hooks/use-actions';
-
-interface ActionBarProps {
-  id: string;
-}
-
-const ActionBar: React.FC<ActionBarProps> = ({ id }) => {
-  const { moveCell, deleteCell } = useActions();
-  return (
-    // todo: refactor these action buttons into separate reusable components, provide onClick function and
-    // specific icon class name as props
-    <div className="action-bar">
-      <button
-        className="button is-primary is-small"
-        onClick={() => moveCell(id, 'up')}
-      >
-        <span className="icon">
-          <i className="fas fa-arrow-up"></i>
-        </span>
-      </button>
-      <button
-        className="button is-primary is-small"
-        onClick={() => moveCell(id, 'down')}
-      >
-        <span className="icon">
-          <i className="fas fa-arrow-down"></i>
-        </span>
-      </button>
-      <button
-        className="button is-primary is-small"
-        onClick={() => deleteCell(id)}
-      >
-        <span className="icon">
-          <i className="fas fa-times"></i>
-        </span>
-      </button>
-    </div>
-  );
-};
-
-export default ActionBar;
