+++ packages/local-client/src/components/code-cell/index.tsx
@@ -1,12 +1,12 @@
 import './code-cell.css';
 import { useEffect } from 'react';
-import CodeEditor from './code-editor';
-import Preview from './preview';
-import Resizable from './resizable';
-import { Cell } from '../state';
-import { useActions } from '../hooks/use-actions';
-import { useTypedSelector } from '../hooks/use-typed-selector';
-import { useCumulativeCode } from '../hooks/use-cumulative-code';
+import CodeEditor from '../code-editor';
+import Preview from '../preview';
+import Resizable from '../resizable';
+import { Cell } from '../../state';
+import { useActions } from '../../hooks/use-actions';
+import { useTypedSelector } from '../../hooks/use-typed-selector';
+import { useCumulativeCode } from '../../hooks/use-cumulative-code';
 
 interface CodeCellProps {
   cell: Cell;
