+++ packages/local-client/src/components/resizable/index.tsx
