+++ packages/local-client/src/components/resizable/resizable.css
