+++ packages/local-client/src/components/text-editor/index.tsx
@@ -1,8 +1,8 @@
 import './text-editor.css';
 import { useState, useEffect, useRef } from 'react';
 import MDEditor from '@uiw/react-md-editor';
-import { Cell } from '../state';
-import { useActions } from '../hooks/use-actions';
+import { Cell } from '../../state';
+import { useActions } from '../../hooks/use-actions';
 
 interface TextEditorProps {
   cell: Cell;
