+++ packages/local-client/src/state/middlewares/persist-middleware.ts
@@ -17,21 +17,20 @@ export const persistMiddleware = ({
     return (action: Action) => {
       next(action);
 
-      if (
-        [
-          ActionType.MOVE_CELL,
-          ActionType.UPDATE_CELL,
-          ActionType.INSERT_CELL_AFTER,
-          ActionType.DELETE_CELL,
-        ].includes(action.type)
-      ) {
-        // todo: refactor switch statements
-        if (timer) {
-          clearTimeout(timer);
-        }
-        timer = setTimeout(() => {
-          saveCells()(dispatch, getState);
-        }, 250);
+      switch (action.type) {
+        case ActionType.MOVE_CELL:
+        case ActionType.UPDATE_CELL:
+        case ActionType.INSERT_CELL_AFTER:
+        case ActionType.DELETE_CELL:
+          if (timer) {
+            clearTimeout(timer);
+          }
+          timer = setTimeout(() => {
+            saveCells()(dispatch, getState);
+          }, 250);
+          break;
+        default:
+          break;
       }
     };
   };
